There is an issue with vehicles following to closely.
This issue can be resolved by setting the first vehicle in all members of a platoon to the next vehicles. This does fix it, but causes an odd issue with sumo crashing when the original vehicle is taken from the simulation upon reaching the absorption layer of the highway. Not sure why it is doing it or how to fix this directly.